api/requirements.txt
mcp==1.6.0
openai==1.57.0
python-dotenv==1.1.0
pydantic==2.11.2
pydantic-settings==2.8.1
uvicorn==0.34.0
httpx==0.27.0
requests==2.31.0
